+++ clean_python/fastapi/security.py
@@ -1,4 +1,3 @@
-from typing import Annotated
 from typing import Optional
 
 from fastapi import Depends
@@ -9,20 +8,21 @@ from fastapi.security import OAuth2AuthorizationCodeBearer
 from clean_python import PermissionDenied
 from clean_python import Tenant
 from clean_python import User
+from clean_python.oauth2 import BaseTokenVerifier
+from clean_python.oauth2 import NoAuthTokenVerifier
 from clean_python.oauth2 import OAuth2SPAClientSettings
 from clean_python.oauth2 import Token
 from clean_python.oauth2 import TokenVerifier
 from clean_python.oauth2 import TokenVerifierSettings
 
 __all__ = [
-    "verify_token",
-    "requires_token",
-    "requires_user",
-    "requires_tenant",
+    "get_token",
+    "get_user",
+    "get_tenant",
     "RequiresScope",
 ]
 
-verifier: Optional[TokenVerifier] = None
+verifier: Optional[BaseTokenVerifier] = None
 
 
 def clear_verifier() -> None:
@@ -31,36 +31,28 @@ def clear_verifier() -> None:
     verifier = None
 
 
-def set_verifier(settings: TokenVerifierSettings) -> None:
+def set_verifier(settings: Optional[TokenVerifierSettings]) -> None:
     global verifier
 
-    verifier = TokenVerifier(settings=settings)
+    if settings is None:
+        verifier = NoAuthTokenVerifier()
+    else:
+        verifier = TokenVerifier(settings=settings)
 
 
-def verify_token(request: Request) -> Optional[Token]:
+def get_token(request: Request) -> Token:
     """A fastapi 'dependable' yielding the validated token"""
     global verifier
 
-    if verifier is None:
-        return None
-
+    assert verifier is not None
     return verifier(request.headers.get("Authorization"))
 
 
-async def requires_token(
-    token: Annotated[Optional[Token], Depends(verify_token)]
-) -> Token:
-    """A fastapi 'dependable' yielding the validated token"""
-    if token is None:
-        raise PermissionDenied("this operation requires a token")
-    return token
-
-
-async def requires_user(token: Annotated[Token, Depends(requires_token)]) -> User:
+async def get_user(token: Token = Depends(get_token)) -> User:
     return token.user
 
 
-async def requires_tenant(token: Annotated[Token, Depends(requires_token)]) -> Tenant:
+async def get_tenant(token: Token = Depends(get_token)) -> Tenant:
     if token.tenant is None:
         raise PermissionDenied("this operation requires a tenant-scoped token")
     return token.tenant
@@ -71,7 +63,7 @@ class RequiresScope:
         assert scope.replace(" ", "") == scope, "spaces are not allowed in a scope"
         self.scope = scope
 
-    async def __call__(self, token: Annotated[Token, Depends(requires_token)]) -> None:
+    async def __call__(self, token: Token = Depends(get_token)) -> None:
         if self.scope not in token.scope:
             raise PermissionDenied(f"this operation requires '{self.scope}' scope")
 
@@ -90,9 +82,7 @@ class OAuth2SPAClientSchema(OAuth2AuthorizationCodeBearer):
             tokenUrl=str(client.token_url),
         )
 
-    async def __call__(
-        self, token: Annotated[Optional[Token], Depends(verify_token)]
-    ) -> None:
+    async def __call__(self, token: Optional[Token] = Depends(get_token)) -> None:
         pass
 
 
@@ -105,7 +95,5 @@ class JWTBearerTokenSchema(HTTPBearer):
     def __init__(self):
         super().__init__(scheme_name="JWT Bearer token", bearerFormat="JWT")
 
-    async def __call__(
-        self, token: Annotated[Optional[Token], Depends(verify_token)]
-    ) -> None:
+    async def __call__(self, token: Optional[Token] = Depends(get_token)) -> None:
         pass